+++ jni/src/simd/similarity_function/default_simd_similarity_function.cpp
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 #include <algorithm>
 #include <array>
 #include <cstddef>